CU-3j8gkyv - Add SP Balance Check Job
