--- src/config/index.ts
@@ -1,5 +1,0 @@
-import sidooh from './sidooh';
-
-export const CONFIG = {
-    sidooh
-}
